+++ RELEASE-NOTES.md
@@ -21,6 +21,7 @@ Picocli follows [semantic versioning](http://semver.org/).
 ## <a name="3.9.2-fixes"></a> Fixed issues
 - [#602] Make CommandLine in AutoComplete use correct IFactory implementation. Thanks to [Mikołaj Krzyżanowski](https://github.com/MikolajK) for the pull request.
 - [#608] Bugfix: defaulting `usageHelp` or `versionHelp` options incorrectly prevents validation of required options and positional parameters. Thanks to [Pietro Braione](https://github.com/pietrobraione) for the bug report.
+- [#612] Bugfix: Usage help for booleans options with `arity = "1"` now correctly show the option parameter in synopsis. Thanks to [prewersk](https://github.com/prewersk) for the bug report.
 - [#606] Doc: Added subcommand example. Thanks to [Andreas Deininger](https://github.com/deining) for the pull request.
 - [#610] Build: add JDKs to Travis CI build.
 - [#609] Created mailing list `picocli at googlegroups dot com`: [picocli Google group](https://groups.google.com/d/forum/picocli).
@@ -29,7 +30,7 @@ Picocli follows [semantic versioning](http://semver.org/).
 No features were deprecated in this release.
 
 ## <a name="3.9.2-breaking-changes"></a> Potential breaking changes
-
+This release has no breaking changes.
 
 
 # <a name="3.9.1"></a> Picocli 3.9.1